fix: net/profile imports and guards; gate SECURITY_CAPABILITIES import under tracing; correct HANDLE conversion
Context:
* Address CI errors: lingering LocalFreeGuard in net.rs; unused imports in profile.rs; HANDLE construction and tracing-only type import.

Change:
* net: use ffi::LocalAllocGuard for current list + target psid; annotate Option type.
* profile: remove unused PCWSTR/CoTaskMemFree imports; use WideString::as_pcwstr; adjust PWSTR imports.
* handles: construct HANDLE from raw pointer instead of isize.
* launch: gate SECURITY_CAPABILITIES import behind feature=tracing.

Risk:
* Low; compile-time fixes only.

Verification:
* Should clear MSRV Windows errors for unused imports and type mismatches.

+++ src/ffi/handles.rs
@@ -40,7 +40,7 @@ impl Handle {
     }
 
     pub(crate) fn as_win32(&self) -> HANDLE {
-        HANDLE(self.as_borrowed().as_raw_handle() as isize)
+        HANDLE(self.as_borrowed().as_raw_handle())
     }
 
     pub(crate) fn into_file(self) -> std::fs::File {
